requirements.txt
bronkhorst-propar==1.1.1
openpyxl==3.2.0b1
pandas==2.2.3
pyserial==3.5